# bin/Debug/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Account & Document Integration API (Nintex Compatible)",
    "description": "REST API for Nintex Workflows to query Customer Accounts and Document Signatures with Base64 encoded file contents.",
    "version": "v1"
  },
  "host": "kyc.runasp.net",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/Accounts/{accountNo}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Retrieves Account and Customer information for a given Account Number.",
        "operationId": "GetAccountByNo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accountNo",
            "in": "path",
            "description": "The Account Number to look up (e.g. ACC1001)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AccountDto"
            }
          }
        }
      }
    },
    "/accounts/{accountNo}/documents/signature": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Retrieves the Signature document for the specified Account Number with Base64-encoded file content for Nintex mapping.",
        "operationId": "GetSignatureDocument",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accountNo",
            "in": "path",
            "description": "The Account Number associated with the signature document (e.g. ACC1001)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DocumentSignatureDto"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "AccountDto": {
      "type": "object",
      "properties": {
        "accountNo": {
          "type": "string",
          "description": "Unique Account Number"
        },
        "customerId": {
          "type": "string",
          "description": "Customer Identifier"
        },
        "customerName": {
          "type": "string",
          "description": "Full Name of the Customer"
        },
        "accountType": {
          "type": "string",
          "description": "Type of Account (e.g. Savings, Checking, Corporate)"
        },
        "accountStatus": {
          "type": "string",
          "description": "Current status of the Account (e.g. Active, Pending, Suspended)"
        },
        "branch": {
          "type": "string",
          "description": "Branch name where account is registered"
        }
      }
    },
    "DocumentSignatureDto": {
      "type": "object",
      "properties": {
        "documentId": {
          "type": "integer",
          "format": "int32",
          "description": "Unique Document ID"
        },
        "accountNo": {
          "type": "string",
          "description": "Account Number associated with this signature document"
        },
        "documentType": {
          "type": "string",
          "description": "Type of document (e.g. Signature)"
        },
        "documentName": {
          "type": "string",
          "description": "Document Name"
        },
        "documentUrl": {
          "type": "string",
          "description": "External or Storage URL of the document"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time",
          "description": "Timestamp when document was created"
        },
        "fileDataBase64": {
          "type": "string",
          "description": "Base64 encoded file attachment content wrapped inside standard JSON object for Nintex database buffer processing"
        },
        "fileName": {
          "type": "string",
          "description": "Original filename for Nintex binary file creation"
        },
        "contentType": {
          "type": "string",
          "description": "MIME Content Type (e.g. image/png, application/pdf)"
        }
      }
    }
  }
}